+++ src/querybuilder/components/QueryPatternsModal.test.tsx
@@ -10,7 +10,6 @@ import { QueryPatternsModal } from './QueryPatternsModal';
 // don't care about interaction tracking in our unit tests
 jest.mock('@grafana/runtime', () => ({
   ...jest.requireActual('@grafana/runtime'),
-  reportInteraction: jest.fn(),
 }));
 
 function getDefaultProps() {