+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 13.1.0
+
+- Persist disabled operations in Loki query builder ([#139](https://github.com/grafana/grafana-loki-datasource/pull/139))
+- Bump @grafana/\* to 13.1.1 and update dependencies ([#137](https://github.com/grafana/grafana-loki-datasource/pull/137))
+
+## 13.0.2
+
+- Streaming key updates ([#135](https://github.com/grafana/grafana-loki-datasource/pull/135))
+- Remove unused Rudderstack events ([#130](https://github.com/grafana/grafana-loki-datasource/pull/130))
+- Update backend dependencies ([#120](https://github.com/grafana/grafana-loki-datasource/pull/120))
+- Update frontend dependencies ([#122](https://github.com/grafana/grafana-loki-datasource/pull/122))
+- Make grafanaDependency prerelease-inclusive ([#126](https://github.com/grafana/grafana-loki-datasource/pull/126))
+
+## 13.0.1
+
+- Bump grafana-plugin-sdk-go v0.292.2 ([#124](https://github.com/grafana/grafana-loki-datasource/pull/124))
+- Pin frontend dependencies ([#121](https://github.com/grafana/grafana-loki-datasource/pull/121))
+- Sync with grafana/grafana ([#115](https://github.com/grafana/grafana-loki-datasource/pull/115))
+
 ## 13.0.0
 
-Initial public release.
+- Initial public release.